+++ StudentManagementSystem/src/main/java/com/yoanpetrov/studentmanagementsystem/repository/StudentCourseEnrollmentRepository.java
@@ -4,5 +4,5 @@ import com.yoanpetrov.studentmanagementsystem.model.StudentCourseEnrollment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StudentCourseEnrollmentRepository
-        extends JpaRepository<StudentCourseEnrollment, Integer> {
+        extends JpaRepository<StudentCourseEnrollment, Long> {
 }
